--- a/.changeset/five-kids-sit.md
+++ b/.changeset/five-kids-sit.md
@@ -1,0 +1,7 @@
+---
+'@ice/rspack-config': patch
+'@ice/shared-config': patch
+'@ice/app': patch
+---
+
+feat: support optimize package import