--- a/.changeset/thin-ligers-taste.md
+++ b/.changeset/thin-ligers-taste.md
@@ -1,0 +1,5 @@
+---
+'@ice/bundles': patch
+---
+
+fix: update binding version for optimize imports