--- a/examples/with-antd5/ice.config.mts
+++ b/examples/with-antd5/ice.config.mts
@@ -2,4 +2,7 @@ import { defineConfig } from '@ice/app';
 
 export default defineConfig(() => ({
   ssg: false,
+  optimization: {
+    optimizePackageImport: true,
+  }
 }));