--- a/packages/shared-config/src/types.ts
+++ b/packages/shared-config/src/types.ts
@@ -205,4 +205,6 @@ export interface Config {
   useDevServer?: boolean;
 
   useDataLoader?: boolean;
+
+  optimizePackageImports?: string[];
 }