Relate to https://github.com/ice-lab/icepack/pull/30

- [x] update binding version to support optimize package import
- [x] support both build-in and user provided packages to be optimized

在简单使用 antd 的项目，解析模块数量从 `2738` 下降到 `309`，启动时间优化从 `358ms` 到 `166ms`

![image](https://github.com/alibaba/ice/assets/4219965/03b8e44b-0c58-4082-b5f2-076d22af45eb)

![image](https://github.com/alibaba/ice/assets/4219965/f8d3b542-862e-4a02-bdd3-5eaa204aebef)


由于减小的压缩阶段的 treeshaking 处理，让整体 build 提升更加明显，从 `1300ms+` 提升值 `300ms+`

![image](https://github.com/alibaba/ice/assets/4219965/d40ab344-a68d-4f16-be13-9447ffb99bbb)
![image](https://github.com/alibaba/ice/assets/4219965/766939d6-b24c-4397-844e-da88805f40a1)